`GotLabel` moved to its own file.
