+++ MVVM-C/Controllers/Article Details/DetailViewVC.swift
@@ -112,33 +112,6 @@ final class DetailViewVC: UIViewController {
 }
 
 
-class GoTLabel: UILabel {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    convenience init(textAlignment: NSTextAlignment, fontSize: CGFloat) {
-        self.init(frame: .zero)
-        self.textAlignment = textAlignment
-        self.font = UIFont.systemFont(ofSize: fontSize, weight: .bold)
-    }
-    
-    func configure() {
-        textColor = .label
-        translatesAutoresizingMaskIntoConstraints = false
-        lineBreakMode = .byTruncatingTail
-        adjustsFontForContentSizeCategory = true
-        numberOfLines = 4
-    }
-}
-
-
 class GoTImageThumbnail: UIImageView {
     
     override init(frame: CGRect) {
