+++ MVVM-C/Model/Article.swift
@@ -2,7 +2,7 @@
 //  ArticleModel.swift
 //  MVVM-C
 //
-//  Created by Usemobile on 09/06/20.
+//  Created by Tulio Parreiras on 09/06/20.
 //  Copyright © 2020 Jakub Homik. All rights reserved.
 //
